Assignment-4/Assignments 01/online_class_projects/Project_3_QR_code_encoder_decoder_Python_Project/requirements.txt
streamlit
qrcode[pil]
numpy
pillow
opencv-python-headless